+++ src/main/java/game/models/abilities/BuffAbility.java
@@ -15,7 +15,7 @@ public class BuffAbility extends Ability{
     private int durationTime;
     private Ability buffedAbility;
 
-    public BuffAbility(double BASE_STRENGTH, String name, AbilityType type, String description, String image, int level, int price, double strength, int cooldownTime, int durationTime) {
+    public BuffAbility(int BASE_STRENGTH, String name, AbilityType type, String description, String image, int level, int price, int strength, int cooldownTime, int durationTime) {
         super(BASE_STRENGTH, name, type, description, image, level, price, strength, cooldownTime);
         this.durationTime = durationTime;
     }
